backend/pyproject.toml
[project]
name = "apilens"
version = "1.0.0"
description = "apilens - API Observability Platform"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "apilens Team" }
]

dependencies = [
    "django>=5.0,<6.0",
    "django-ninja>=1.0,<2.0",
    "django-cors-headers>=4.3,<5.0",
    "psycopg[binary]>=3.1,<4.0",
    "python-dotenv>=1.0,<2.0",
    "gunicorn>=21.0,<23.0",
    "httpx>=0.25,<1.0",
    "authlib>=1.6.6",
    "PyJWT[crypto]>=2.8,<3.0",
]
